requirements.txt
langchain
langchain-core
langchain-openai
openevals
playwright
pydantic
python-dotenv
langsmith

